LayoutTests/platform/mac/accessibility/mac/selected-text-range-unconnected-object-expected.txt

This tests that the selected text marker range is valid when it references an unconnected node that changes.

Text Marker: [Length: 5]
	START: AXRole: AXStaticText
	END: AXRole: AXStaticText

Text Marker Range Attributed Text: AXFont - {
    AXFontFamily = Helvetica;
    AXFontName = Helvetica;
    AXFontSize = 11;
}, Hello

Inserting 'world'
Text Marker: [Length: 11]
	START: AXRole: AXStaticText
	END: AXRole: AXStaticText

Text Marker Range Attributed Text: AXFont - {
    AXFontFamily = Helvetica;
    AXFontName = Helvetica;
    AXFontSize = 11;
}, Hello world


PASS successfullyParsed is true

TEST COMPLETE
